Changed normalisation routine LALInspiralWaveNormaliseLSO so that the signal normal is computed by using power only up to fCutoff=fFinal. Also, the Fourier domain signal is terminated beyond fCutoff=fFinal. Original: 77c4adeaff45e1b110e929ca62a07d797e6d123c

+++ lal/packages/noisemodels/src/LALRandomInspiralSignal.c
@@ -177,8 +177,8 @@ LALRandomInspiralSignal
 	   valid = 0;
 	   while (!valid) 
 	   {
-		   epsilon1 = random()/(float)RAND_MAX;
-		   epsilon2 = random()/(float)RAND_MAX;
+		   epsilon1 = (float) random()/(float)RAND_MAX;
+		   epsilon2 = (float) random()/(float)RAND_MAX;
 		   switch (randIn->param.massChoice) 
 		   {
 			   case m1Andm2:
